Last Update 14/may

+++ src/Application/Employees/Commands/RemoveEmployee/RemoveEmployeeCommand.cs
@@ -1,4 +1,5 @@
 ﻿using Employees.src.Application.Common.Interfaces;
+using Employees.src.Application.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -24,6 +25,12 @@ namespace Employees.src.Application.Employees.Commands.RemoveEmployee
         public async Task<Unit> Handle(RemoveEmployeeCommand request, CancellationToken cancellationToken)
         {
             var employee = await _context.Employees.FindAsync(request.Id);
+
+            if (employee == null)
+            {
+                throw new NotFoundException(nameof(employee), request.Id);
+            }
+
             _context.Employees.Remove(employee);
             await _context.SaveChangesAsync(cancellationToken);
